# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BéninExpense AI - Plateforme de gestion des dépenses et conformité fiscale
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: accountId
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
            format: date
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, validated, rejected, synced]
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"

    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    get:
      operationId: getExpense
      tags: [expenses]
      summary: Get expense by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Not found

    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /expenses/export:
    get:
      operationId: exportExpenses
      tags: [expenses]
      summary: Export expenses as CSV
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, validated, rejected, synced]
        - name: category
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string
  /expenses/{id}/validate:
    post:
      operationId: validateExpense
      tags: [expenses]
      summary: Validate an expense (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}/reject:
    post:
      operationId: rejectExpense
      tags: [expenses]
      summary: Reject an expense (admin/accountant)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # AI Assistant
...
components:
  schemas:
...
    Expense:
      type: object
      required: [id, description, amount, category, status, createdAt]
      properties:
        id:
          type: integer
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        status:
          type: string
          enum: [pending, validated, rejected, synced]
        accountId:
          type: ["integer", "null"]
        employeeId:
          type: ["integer", "null"]
        employeeName:
          type: ["string", "null"]
        receiptUrl:
          type: ["string", "null"]
        dgiQrCode:
          type: ["string", "null"]
        dgiReference:
          type: ["string", "null"]
        dgiStatus:
          type: string
          enum: [not_submitted, pending, normalized, failed]
        notes:
          type: ["string", "null"]
        submittedBy:
          type: ["integer", "null"]
          description: ID of the user who submitted the expense
        rejectionReason:
          type: ["string", "null"]
          description: Reason for rejection if status is rejected
        riskLevel:
          type: string
          enum: [none, low, medium, high]
          default: none
        flagReason:
          type: ["string", "null"]
          description: Reason for flagging the expense as suspicious
        createdAt:
          type: string
          format: date-time
    ExpenseInput:
      type: object
      required: [description, amount, category]
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        accountId:
          type: integer
        employeeId:
          type: integer
        notes:
          type: string
        receiptUrl:
          type: string
    ExpenseUpdate:
      type: object
      properties:
        description:
          type: string
        amount:
          type: number
        category:
          type: string
        status:
          type: string
          enum: [pending, validated, rejected]
        notes:
          type: string

    # Budget schemas
...